Project 2 Email otp verification/requirement.txt

asgiref==3.5.1
auditlog3==1.0.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pyotp==2.6.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
django-summernote==0.8.20.0
django-rest-swagger==2.2.0
boto3==1.22.8
botocore==1.25.8
django-storages==1.12.3
drf-yasg==1.20.0
Pillow==9.1.0
google-api-core==2.8.1
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.2